+++ app/Console/Commands/EsIndexDocuments.php
@@ -20,12 +20,13 @@
 
 namespace App\Console\Commands;
 
+use App\Libraries\Elasticsearch\Indexing;
 use App\Models\Beatmapset;
 use App\Models\Forum\Post;
 use App\Models\User;
 use Illuminate\Console\Command;
 
-class EsIndexDocuments extends EsIndexCommand
+class EsIndexDocuments extends Command
 {
     const ALLOWED_TYPES = [
         'beatmapsets' => Beatmapset::class,
@@ -47,9 +48,95 @@ class EsIndexDocuments extends EsIndexCommand
      */
     protected $description = 'Indexes documents into Elasticsearch.';
 
+    protected $cleanup;
+    protected $inplace;
+    protected $suffix;
+    protected $types;
+    protected $yes;
+
+    /**
+     * Execute the console command.
+     *
+     * @return mixed
+     */
+    public function handle()
+    {
+        $this->readOptions();
+        $this->suffix = !$this->inplace ? '_'.time() : '';
+
+        $oldIndices = [];
+        foreach ($this->types as $type) {
+            $oldIndices[] = Indexing::getOldIndices($type::esIndexName());
+        }
+
+        $oldIndices = array_flatten($oldIndices);
+
+        $continue = $this->starterMessage($oldIndices);
+        if (!$continue) {
+            return $this->error('User aborted!');
+        }
+
+        $start = time();
+
+        $indices = $this->index();
+
+        $this->finish($indices, $oldIndices);
+        $this->warn("\nIndexing completed in ".(time() - $start).'s');
+    }
+
+    protected function finish(array $indices, array $oldIndices)
+    {
+        if (!$this->inplace && $this->cleanup) {
+            foreach ($oldIndices as $index) {
+                $this->warn("Removing '{$index}'...");
+                Indexing::deleteIndex($index);
+            }
+        }
+    }
+
+    /**
+     * Indexes and returns the names of the indices.
+     *
+     * @return array names of the indices indexed to.
+     */
+    protected function index()
+    {
+        $indices = [];
+        foreach ($this->types as $type) {
+            $count = $type::esIndexingQuery()->count();
+            $bar = $this->output->createProgressBar($count);
+
+            if (!$this->inplace) {
+                $indexName = "{$type::esIndexName()}{$this->suffix}";
+
+                $this->info("Indexing {$type} into {$indexName}");
+
+                $type::esIndexIntoNew(1000, $indexName, function ($progress) use ($bar) {
+                    $bar->setProgress($progress);
+                });
+
+                $indices[] = $indexName;
+            } else {
+                $this->info("In-place indexing {$type} into {$type::esIndexName()}");
+                $type::esReindexAll(1000, 0, [], function ($progress) use ($bar) {
+                    $bar->setProgress($progress);
+                });
+
+                $indices[] = $type::esIndexName();
+            }
+
+            $bar->finish();
+            $this->line("\n");
+        }
+
+        return $indices;
+    }
+
     protected function readOptions()
     {
-        parent::readOptions();
+        $this->inplace = $this->option('inplace');
+        $this->cleanup = $this->option('cleanup');
+        $this->yes = $this->option('yes');
 
         if ($this->option('types')) {
             $types = explode(',', $this->option('types'));
@@ -63,5 +150,26 @@ class EsIndexDocuments extends EsIndexCommand
         } else {
             $this->types = array_values(static::ALLOWED_TYPES);
         }
+    }
+
+    protected function starterMessage(array $oldIndices)
+    {
+        if ($this->inplace) {
+            $this->warn('Running in-place reindex.');
+            $confirmMessage = 'This will reindex in-place (schemas must match)';
+        } else {
+            $this->warn('Running index transfer.');
+
+            if ($this->cleanup) {
+                $this->warn(
+                    "The following indices will be deleted on completion!\n"
+                    .implode("\n", $oldIndices)
+                );
+            }
+
+            $confirmMessage = 'This will create new indices';
+        }
+
+        return $this->yes || $this->confirm("{$confirmMessage}, begin indexing?");
     }
 }
